feat(US-054): 工数/稼働率を小数3位・進捗率0.1%精度＋入力枠幅とスピン刻み
- Task.progress / ProgressReport.progress を Float 化(0.1%単位)。SQLite 動的型で既存DB互換
- 進捗バリデーション(tasks/dailyReports)を int 制約から数値へ。expectedProgress/clamp/
  集計を 0.1% 丸めに。overallProgress も 0.1% 丸め
- 入力 step: 工数/稼働率=0.001、進捗=0.1。ガント/WBS編集/見積確認の数値欄と列幅を拡大
- backend 82 / frontend 45 tests green

Closes #122

+++ backend/src/routes/dailyReports.ts
@@ -39,7 +39,7 @@ const createInput = z.object({
     .array(
       z.object({
         taskId: z.string().min(1),
-        progress: z.number().int().min(0).max(100),
+        progress: z.number().min(0).max(100),
         comment: z.string().optional(),
       }),
     )
